+++ packages/core/shared/src/index.ts
@@ -1,1 +1,3 @@
-export * from "./types.js";
+export * from "./loglayer.types.js";
+export * from "./plugin.types.js";
+export * from "./common.types.js";